+++ src/main/java/com/caffeine/gwanghwamun/domain/user/controller/AuthController.java
@@ -3,7 +3,7 @@ package com.caffeine.gwanghwamun.domain.user.controller;
 import com.caffeine.gwanghwamun.common.response.ApiResponse;
 import com.caffeine.gwanghwamun.common.response.ResponseUtil;
 import com.caffeine.gwanghwamun.common.success.SuccessCode;
-import com.caffeine.gwanghwamun.domain.user.dto.SignupReqDTO;
+import com.caffeine.gwanghwamun.domain.user.dto.request.SignupReqDTO;
 import com.caffeine.gwanghwamun.domain.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
