+++ src/main/java/com/caffeine/gwanghwamun/domain/user/dto/request/LoginReqDTO.java
@@ -1,4 +1,4 @@
-package com.caffeine.gwanghwamun.domain.user.dto;
+package com.caffeine.gwanghwamun.domain.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
