+++ src/main/java/com/caffeine/gwanghwamun/domain/user/dto/request/SignupReqDTO.java
@@ -1,4 +1,4 @@
-package com.caffeine.gwanghwamun.domain.user.dto;
+package com.caffeine.gwanghwamun.domain.user.dto.request;
 
 import com.caffeine.gwanghwamun.domain.user.entity.User;
 import com.caffeine.gwanghwamun.domain.user.entity.UserRoleEnum;
